% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/envvar_utility_functions.R
\name{env_or_default}
\alias{env_or_default}
\title{Return the value of the named environment variable(s), or a default value if NA}
\usage{
env_or_default(envvar_names, values_if_missing)
}
\arguments{
\item{envvar_names}{A vector of envrionment variable names to retrieve}

\item{values_if_missing}{A vector of strings, matched by position, to use as defaults}
}
\value{
A vector (named if length > 1) of a matching length (> 0) with the conditioned values
}
\description{
The common use-case for this is when it is expected to setup execution-specific
variable values from a calling environment but if the script is run in a
standalone/testing mode that it can make use of the defaults to still
operate in a controlled manner.
}
\details{
While bulk lookups can be accomplished readability will be hampered if
extremely long vectors are used.  In that situation multiple single-pair
calls is recommended.  This also has the beneficial property of not attaching
names to the resultant vector which makes variable assignment easier.

Wraps a call to Sys.getenv
}
\examples{
env_or_default("UNKNOWN", "DEFAULT")
env_or_default(c("HOME", "UNKNOWN"), c("/tmp", "DEFAULT"))
}

